% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allCBRTCategories}
\alias{allCBRTCategories}
\title{Data categories}
\format{
A data table with 23 rows and 2 variables:
\describe{
  \item{cid}{Category ID}
  \item{topic}{Topic}
}
}
\usage{
allCBRTCategories
}
\description{
A dataset containing the list of all data categories in the CBRT dataset
}
\keyword{datasets}
