% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allCBRTGroups}
\alias{allCBRTGroups}
\title{Data groups}
\format{
A data table with 323 rows and 10 variables:
\describe{
  \item{cid}{Category ID}
  \item{groupCode}{Data group code (used for downloading the data)}
  \item{groupName}{Name of the data group}
  \item{freq}{Time series frequency code}
  \item{source}{Data source}
  \item{sourceLink}{URL of the data source}
  \item{revisionPolicy}{Revision policy for the data group}
  \item{firstDate}{The beginning date for the data}
  \item{lastDate}{The end date for the data}
  \item{appLing}{URL}
}
}
\usage{
allCBRTGroups
}
\description{
A dataset containing the list of all data groups in the CBRT dataset
}
\keyword{datasets}
