% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allCBRTSeries}
\alias{allCBRTSeries}
\title{Data series}
\format{
A data table frame with 22,243 rows and 12 variables:
\describe{
  \item{cid}{Category ID}
  \item{topic}{Topic}
  \item{groupCode}{Data group code}
  \item{groupName}{Name of the data group}
  \item{freq}{Time series frequency code}
  \item{seriesCode}{Data series code (used for downloading the data)}
  \item{seriesName}{Name of the data series}
  \item{start}{Starting date, DD-MM-YYYY}
  \item{end}{Ending date, DD-MM-YYYY}
  \item{aggMethod}{Data aggregation method}
  \item{freqname}{Time series frequency}
  \item{tag}{Tages (keywords)}
}
}
\usage{
allCBRTSeries
}
\description{
A dataset containing the list of all data series in the CBRT dataset
}
\keyword{datasets}
