% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbRt-package.R
\docType{package}
\name{cbRt-package}
\alias{cbRt-package}
\title{cbRt: R Interface to EDDS of CBRT}
\description{
An R interface to the Electronic Data Delivery System (EVDS) of the
Central Bank of the Republic of Turkey (CBRT). Provides functions to
retrieve economic data and metadata from the EVDS API v3, with automatic
handling of the 150 observation limit through intelligent chunking based
on series frequency detection.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emraher/cbRt}
  \item \url{https://eremrah.com/cbRt/}
  \item Report bugs at \url{https://github.com/emraher/cbRt/issues}
}

}
\author{
\strong{Maintainer}: Emrah Er \email{eer@eremrah.com} (\href{https://orcid.org/0000-0001-9909-7479}{ORCID})

}
\keyword{internal}
