% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBRT.R
\name{CBRT_fread}
\alias{CBRT_fread}
\title{CBT API function}
\usage{
CBRT_fread(fileName, CBRTKey)
}
\arguments{
\item{fileName}{URL for API}

\item{CBRTKey}{Individual key for CB API}
}
\value{
The data from CBRT
}
\description{
Reading the data from the CBT server
}
\keyword{internal}
