% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbRt_meta.R
\name{cbrt_meta}
\alias{cbrt_meta}
\title{Download information for all available series}
\usage{
cbrt_meta(token = NULL)
}
\arguments{
\item{token}{API Key}
}
\value{
A data frame with metadata for all available series
}
\description{
Download information for all available series
}
\examples{
# Download information for all available series
\dontrun{cbrt_meta(token = Sys.getenv("EVDS_TOKEN"))}

}
