% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbRt_helper.R
\name{cbRt_url}
\alias{cbRt_url}
\title{Helper functions}
\usage{
cbRt_url(series, startDate, endDate, token, aggregationTypes = NULL,
  formulas = NULL, freq = NULL)
}
\description{
Helper functions
}
\keyword{internal}
