% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CBRTagg}
\alias{CBRTagg}
\title{Aggregation methods}
\format{
A data table frame with 6 rows and 2 variables:
\describe{
  \item{Code}{Aggregation code}
  \item{Aggregation}{Aggregation method}
}
}
\usage{
CBRTagg
}
\description{
A dataset containing the list of aggregation methods
}
\keyword{datasets}
