% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CBRTfreq}
\alias{CBRTfreq}
\title{Frequenciess}
\format{
A data table frame with 8 rows and 4 variables:
\describe{
  \item{freq}{Frequency code}
  \item{tfreq}{Frequency code (internal use)}
  \item{FreqEng}{Frequency name (English)}
  \item{FreqTr}{Frequency name (Turkish)}
}
}
\usage{
CBRTfreq
}
\description{
A dataset containing the list of frequencies
}
\keyword{datasets}
