% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBRT.R
\name{changeASCII}
\alias{changeASCII}
\title{Changing characters to ASCII}
\usage{
changeASCII(x)
}
\arguments{
\item{x}{String to be modifies}
}
\value{
Modified string
}
\description{
Changes non-ASCII characters to ASCII
}
\keyword{internal}
