% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBRT.R
\name{formatTime}
\alias{formatTime}
\title{Formatting time series}
\usage{
formatTime(x)
}
\arguments{
\item{x}{a data series}
}
\value{
formatted object
}
\description{
Sets the format of a time series object retrieved from the CBRT database
}
\keyword{internal}
