% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBRT.R
\name{getAllGroupsInfo}
\alias{getAllGroupsInfo}
\title{Information about data groups}
\usage{
getAllGroupsInfo(CBRTKey = myCBRTKey)
}
\arguments{
\item{CBRTKey}{Your personal CBRT access key}
}
\value{
a data.table object
}
\description{
Creates a meta data object for all data gorups
}
\examples{
\dontrun{
allCBRTGroups <- getAllGroupsInfo()
}

}
