% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBRT.R
\name{getAllSeriesInfo}
\alias{getAllSeriesInfo}
\title{Information about data series}
\usage{
getAllSeriesInfo(CBRTkey = myCBRTKey)
}
\arguments{
\item{CBRTkey}{Your personal CBRT access key}
}
\value{
a data.table object
}
\description{
Creates a meta data object for all data series
}
\examples{
allCBRTSeries <- getAllSeriesInfo()

}
