% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBRT.R
\name{getDataGroup}
\alias{getDataGroup}
\title{Downloading data groups}
\usage{
getDataGroup(group, CBRTkey = myCBRTKey, freq,
  startDate = "01-01-1950", endDate, na.rm = T)
}
\arguments{
\item{group}{Code for the data group}

\item{CBRTkey}{Your personal CBRT access key}

\item{freq}{Numeric, the frequency of the data series}

\item{startDate}{The beginning date for data series (DD-MM-YYYY)}

\item{endDate}{The ending date for data series (DD-MM-YYYY)}

\item{na.rm}{Logical variable to drop all missing dates}
}
\value{
a data.table object
}
\description{
Downloads all data series of a data group
}
\examples{
myData <- getDataGroup("bie_dbafod")

}
