% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBRT.R
\name{getDataSeries}
\alias{getDataSeries}
\title{Downloading data series}
\usage{
getDataSeries(series, CBRTkey = myCBRTKey, freq, aggType,
  startDate = "01-01-1950", endDate, na.rm = T)
}
\arguments{
\item{series}{A vector of data series' codes}

\item{CBRTkey}{Your personal CBRT access key}

\item{freq}{Numeric, the frequency of the data series}

\item{aggType}{Aggregation of data series}

\item{startDate}{The beginning date for data series (DD-MM-YYYY)}

\item{endDate}{The ending date for data series (DD-MM-YYYY)}

\item{na.rm}{Logical variable to drop all missing dates}
}
\value{
a data.table object
}
\description{
Downloads one or more data series from the CBRT datasets
}
\examples{
mySeries <- getDataSeries("TP.D1TOP")
mySeries <- getDataSeries(c("TP.D1TOP", "TP.D2HAZ", "TP.D4TCMB"))
mySeries <- getDataSeries(c("TP.D1TOP", "TP.D2HAZ", "TP.D4TCMB", startDate="01-01-2010"))

}
