% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_ohlc_single.R
\name{image_ohlc_single}
\alias{image_ohlc_single}
\title{Create an image out of two pre-processed (numeric) time-series}
\usage{
image_ohlc_single(
  ohlc,
  type = "line",
  path = tempdir(),
  filename = "img",
  width = 200,
  height = 200,
  colors = c("black", "red", "green", "lightgrey")
)
}
\arguments{
\item{ohlc}{ohlc xts of time series}

\item{path}{path where to save the image to}

\item{filename}{file name of the image}

\item{width}{the width of the image in pixels}

\item{height}{the height of the image in pixels}

\item{colors}{a vector of colors to be used in sequence (currently 4: dark, up, down, light)}
}
\value{
currently TRUE in every case
}
\description{
\code{image_ohlc_single} creates an image for chart-based Deep Learning
}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
