% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBRT.R
\name{searchCBRT}
\alias{searchCBRT}
\title{Variable search}
\usage{
searchCBRT(keywords, field = c("groups", "categories", "series"), tags = FALSE)
}
\arguments{
\item{keywords}{A vector of keywords}

\item{field}{The name of the field to be searched ("groups", "categories" or
"series"). We recommend searching first the "groups" names.}

\item{tags}{A logical variable that indicates if the tags to be searched}
}
\value{
a data.table object
}
\description{
Search for keywords in the CBRT datasets
}
\examples{
searchCBRT(c("production", "labor", "labour"))
searchCBRT(c("production", "labor", "labour"), field = "series")
searchCBRT(c("production", "labor", "labour"), tags = TRUE)

}
