% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBRT.R
\name{showGroupInfo}
\alias{showGroupInfo}
\title{Information about a data group}
\usage{
showGroupInfo(gCode, verbose = TRUE)
}
\arguments{
\item{gCode}{the code for the data group}

\item{verbose}{TRUE turns on status and information messages to the console}
}
\value{
a data.table object
}
\description{
Shows information about a data group
}
\examples{
showGroupInfo("bie_apifon")

}
