% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBRT.R
\name{showSeriesNames}
\alias{showSeriesNames}
\title{Showing variable names}
\usage{
showSeriesNames(gCode)
}
\arguments{
\item{gCode}{the code for the data group}
}
\value{
a data.table object
}
\description{
Shows the names of all variables in a data group
}
\examples{
showSeriesNames("bie_apifon")

}
