% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbrt.R
\docType{package}
\name{cbrt-package}
\alias{cbrt}
\alias{cbrt-package}
\title{Chart-based Robo Trading using Artificial Intelligence}
\description{
Chart-based Robo Trading using Artificial Intelligence
}
\references{
\url{https://algorithmic.finance/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://algorithmic.finance/}
}

}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
