% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBRT-package.R
\docType{package}
\name{CBRT-package}
\alias{CBRT}
\alias{CBRT-package}
\title{CBRT: CBRT Data on Turkish Economy}
\description{
The package contains functions to access to the Central Bank of the Republic of Turkey's Electronic Data Delivery System. It also contains the lists of all data categories, data groups and data series for searching available data series. These lists can alco be updated any time by the user. As of January 26, 2019, the dataset includes 26,243 time series.
}
\keyword{internal}
