% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbRt.R
\name{cbRt}
\alias{cbRt}
\title{_PACKAGE
cbRt}
\description{
R Interface to CBRT EDDS
}
\details{
The Electronic Data Delivery System (EDDS) is a dynamic and interactive
system that presents statistical time series data produced by the Central
Bank of the Republic of Turkey (CBRT) and/or data produced by other
institutions and compiled by the CBRT.
Datasets can be accessed through \code{cbRt} functions.
}
