% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbRt_get.R
\name{cbrt_get}
\alias{cbrt_get}
\title{Get the series from CBRT}
\usage{
cbrt_get(
  series,
  start_date,
  end_date,
  nd = TRUE,
  as = c("tibble", "tsibble", "data.frame", "data.table")
)
}
\arguments{
\item{series}{Series Code

\code{series} argument can be obtained from CBRT webpage. Search CBRT
webpage in order to find out the series code and use it in the
\code{series} argument. The serial codes are displayed. If more than
one series is selected, it takes up parameters as the number of
series, the serial codes are separated by a "-" sign.}

\item{start_date}{Start date

\code{start_date} argument is the series start date as dd-mm-yyyy
format. In order to display the frequency of the desired series, the
first day of the corresponding frequency must be stated in the start
date as dd-mm-yyyy format.}

\item{end_date}{End date

\code{end_date} argument is the series end date as dd-mm-yyyy format.}

\item{nd}{Convert ND values to NA

\code{nd} argument is a TRUE or FALSE argument. Data retrieved sometimes
includes ND terms. If \code{nd} is set to TRUE, all NDs are converted to
NAs.}

\item{as}{Type of data to return

Argument \code{as} can be set to either \code{tibble}, \code{tsibble},
\code{data.frame}, or \code{data.table} to obtain different types of data
classes. \code{tibble} is the default output class.}
}
\value{
A data frame (tibble, tsibble, data.frame, or data.table)
}
\description{
Get the series from CBRT
}
\examples{
# Download the given series for the given dates
\dontrun{cbrt_get(series = "TP.DK.USD.A",
                  start_date = "01-01-2017",
                  end_date = "01-01-2018",
                  token = APIkey)}

}
