% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbrt_meta_data}
\alias{cbrt_meta_data}
\title{CBRT EDDS Series Information Data}
\format{
An object of class \code{"tibble"}
}
\usage{
data(cbrt_meta_data)
}
\description{
CBRT EDDS Series Information Data
}
\examples{

\dontrun{data(cbrt_meta_data)
cbrt_meta_data \%>\%
  dplyr::filter(SERIE_CODE \%in\% series) \%>\%
 dplyr::select(SERIE_CODE, FREQUENCY_STR)}
}
\keyword{datasets}
