% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBRT.R
\name{getAllCategoriesInfo}
\alias{getAllCategoriesInfo}
\title{Information about data categories}
\usage{
getAllCategoriesInfo(CBRTKey = myCBRTKey)
}
\arguments{
\item{CBRTKey}{Your personal CBRT access key}
}
\value{
a data.table object
}
\description{
Creates a meta data object for all categories
}
\examples{
\dontrun{
allCBRTCategories <- getAllCategoriesInfo()
}

}
