% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBRT.R
\name{getAllSeriesInfo}
\alias{getAllSeriesInfo}
\title{Information about data series}
\usage{
getAllSeriesInfo(CBRTKey = myCBRTKey, verbose = TRUE)
}
\arguments{
\item{CBRTKey}{Your personal CBRT access key}

\item{verbose}{TRUE turns on status and information messages to the console}
}
\value{
a data.table object
}
\description{
Creates a meta data object for all data series
}
\examples{
\dontrun{
allCBRTSeries <- getAllSeriesInfo()
}

}
