% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.gen.R
\name{image.gen}
\alias{image.gen}
\title{Create an image out of two pre-processed (numeric) time-series}
\usage{
\method{image}{gen}(
  series1,
  series2,
  path = getwd(),
  filename = "img",
  width = 200,
  height = 200,
  colors = c("red", "blue"),
  ...
)
}
\arguments{
\item{series1}{numeric vector of values of time-series 1}

\item{series2}{numeric vector of values of time-series 2}

\item{path}{path where to save the image to}

\item{filename}{file name of the image}
}
\value{
currently TRUE
}
\description{
\code{image.gen} creates an image for
}
\author{
Ronald Hochreiter, \email{ronald@algorithmic.finance}
}
